aml_requirements.txt
# Azure ML
azure-ai-ml
azure-core
azure-identity
azure-mgmt-authorization
jsonargparse
pip-requirements-parser

torch
